Combined service report scripts into one
